requirements.txt
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
transformers==4.48.0
torch==2.5.1
python-dotenv
requests
numpy
